+++ packages/evmApi/src/resolvers/events/getContractLogs.ts
@@ -1,6 +1,6 @@
 import { createPaginatedEndpointFactory, createPaginatedEndpoint, PaginatedParams } from '@moralisweb3/api-utils';
 import { Camelize, toCamelCase } from '@moralisweb3/core';
-import { EvmChainish, EvmAddressish, EvmAddress, EvmTransactionLog } from '@moralisweb3/evm-utils';
+import { EvmChainish, EvmAddressish, EvmAddress, EvmTransactionLog } from '@moralisweb3/common-evm-utils';
 import { operations } from '../../generated/types';
 import { EvmChainResolver } from '../EvmChainResolver';